+++ spam/03_spam_data_slicing_tutorial.py
@@ -37,7 +37,7 @@ pd.set_option("display.max_colwidth", -1)
 
 # %% [markdown]
 # _Note:_ this tutorial differs from the labeling tutorial in that we use ground truth labels in the train split for demo purposes.
-# In practice, data slicing is agnostic to the _training labels_ used as inputs — you can use Snorkel-generated labels as inputs to this pipeline.
+# SFs are intended to be used *after the training set has already been labeled* by LFs (or by hand) in the trainind data pipeline.
 
 # %%
 from utils import load_spam_dataset
@@ -78,17 +78,6 @@ def short_link(x):
 sfs = [short_link]
 
 # %% [markdown]
-# For our $n$ examples and $k$ slices in each split, we apply the SF to our data to create an $n \times k$ matrix. (So far, $k=1$).
-
-# %%
-from snorkel.slicing import PandasSFApplier
-
-applier = PandasSFApplier(sfs)
-S_train = applier.apply(df_train)
-S_valid = applier.apply(df_valid)
-S_test = applier.apply(df_test)
-
-# %% [markdown]
 # ### Visualize slices
 
 # %% [markdown]
@@ -114,25 +103,10 @@ short_link_df[["text", "label"]]
 # First, we'll featurize the data—as you saw in the introductory Spam tutorial, we can extract simple bag of words features and store them as numpy arrays.
 
 # %%
-import torch
 from sklearn.feature_extraction.text import CountVectorizer
+from utils import df_to_torch_features
 
 vectorizer = CountVectorizer(ngram_range=(1, 1))
-
-
-def df_to_torch_features(vectorizer, df, fit_train=False):
-    words = [row.text for i, row in df.iterrows()]
-
-    if fit_train:
-        feats = vectorizer.fit_transform(words)
-    else:
-        feats = vectorizer.transform(words)
-    X = feats.todense()
-    Y = df["label"].values
-    return X, Y
-
-
-# %%
 X_train, Y_train = df_to_torch_features(vectorizer, df_train, fit_train=True)
 X_valid, Y_valid = df_to_torch_features(vectorizer, df_valid, fit_train=False)
 X_test, Y_test = df_to_torch_features(vectorizer, df_test, fit_train=False)
@@ -145,20 +119,9 @@ X_test, Y_test = df_to_torch_features(vectorizer, df_test, fit_train=False)
 # In this task, we'd like to store the bag-of-words `bow_features` in our `X_dict`, and we have one set of labels (for now) correpsonding to the `spam_task`.
 
 # %%
-from snorkel.classification.data import DictDataset, DictDataLoader
+from utils import create_dict_dataloader
 
 BATCH_SIZE = 32
-
-
-def create_dict_dataloader(X, Y, split, **kwargs):
-    """Create a DictDataLoader for bag-of-words features."""
-    ds = DictDataset(
-        name="spam_dataset",
-        split=split,
-        X_dict={"bow_features": torch.FloatTensor(X)},
-        Y_dict={"spam_task": torch.LongTensor(Y)},
-    )
-    return DictDataLoader(ds, **kwargs)
 
 
 dl_train = create_dict_dataloader(
@@ -181,54 +144,15 @@ dl_valid.dataset
 # ### Define [`MultitaskClassifier`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/classification/snorkel.classification.MultitaskClassifier.html)
 #
 # We define a simple Multi-Layer Perceptron (MLP) architecture to learn from the `bow_features`.
+# We do so by initializing a `spam_task` using Snorkel's [`Task`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/classification/snorkel.classification.Task.html) API.
 #
-# *Note: the following might feel like extra steps to define what is a very simple architecture (e.g. `sklearn`'s [`MLPClassifier`](https://scikit-learn.org/stable/modules/generated/sklearn.neural_network.MLPClassifier.html)), but this will lend us additional flexibility later in the pipeline!*
-
-# %% [markdown]
-# To start, we define a `module_pool` with all the PyTorch modules that we'll want to include in our network.
+# *We note that it's certainly possible to define an MLP in a simple framework (e.g. `sklearn`'s [`MLPClassifier`](https://scikit-learn.org/stable/modules/generated/sklearn.neural_network.MLPClassifier.html)), but the multitask API will lend us additional flexibility later in the pipeline!*
 
 # %%
-import torch.nn as nn
+from utils import create_spam_task
 
 bow_dim = X_train.shape[1]
-module_pool = nn.ModuleDict(
-    {
-        "mlp": nn.Sequential(nn.Linear(bow_dim, bow_dim), nn.ReLU()),
-        "prediction_head": nn.Linear(bow_dim, 2),
-    }
-)
-
-# %% [markdown]
-# Then, we specify the desired `task_flow` through each module.
-
-# %%
-from snorkel.classification.task import Operation
-
-task_flow = [
-    Operation(name="input_op", module_name="mlp", inputs=[("_input_", "bow_features")]),
-    Operation(name="head_op", module_name="prediction_head", inputs=[("input_op", 0)]),
-]
-
-# %% [markdown]
-# With these pieces, we're ready to define a [`Task`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/classification/snorkel.classification.Task.html) in Snorkel for spam classification.
-
-# %%
-from functools import partial
-from snorkel.analysis import Scorer
-from snorkel.classification import (
-    Task,
-    cross_entropy_from_outputs,
-    softmax_from_outputs,
-)
-
-spam_task = Task(
-    name="spam_task",
-    module_pool=module_pool,
-    task_flow=task_flow,
-    loss_func=partial(cross_entropy_from_outputs, "head_op"),
-    output_func=partial(softmax_from_outputs, "head_op"),
-    scorer=Scorer(metrics=["accuracy", "f1"]),
-)
+spam_task = create_spam_task(bow_dim)
 
 # %% [markdown]
 # We'll initialize a  [`MultitaskClassifier`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/classification/snorkel.classification.MultitaskClassifier.html) with the `spam_task` we've created, initialize a corresponding [`Trainer`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/classification/snorkel.classification.Trainer.html), and `fit` to our dataloaders!
@@ -277,6 +201,17 @@ df_valid[["text", "label"]].iloc[error_buckets[(1, 0)]].head()
 
 # %% [markdown]
 # In order to monitor performance on our `short_link` slice, we add labels to an existing dataloader.
+# First, for our $n$ examples and $k$ slices in each split, we apply the SF to our data to create an $n \times k$ matrix. (So far, $k=1$).
+
+# %%
+from snorkel.slicing import PandasSFApplier
+
+applier = PandasSFApplier(sfs)
+S_train = applier.apply(df_train)
+S_valid = applier.apply(df_valid)
+S_test = applier.apply(df_test)
+
+# %% [markdown]
 # Specifically, [`add_slice_labels`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/slicing/snorkel.slicing.add_slice_labels.html#snorkel.slicing.add_slice_labels) will add two sets of labels for each slice:
 # * `spam_task_slice:{slice_name}_ind`: an indicator label, which corresponds to the outputs of the slicing functions.
 # These indicate whether each example is in the slice (`label=1`)or not (`label=0`).
@@ -324,30 +259,19 @@ sklearn_model.score(X_test, Y_test)
 
 # %%
 from snorkel.utils import preds_to_probs
-from snorkel.slicing import SliceScorer
-
 
 preds_test = sklearn_model.predict(X_test)
+probs_test = preds_to_probs(preds_test, 2)
+
+# %%
+from snorkel.analysis import Scorer
+from snorkel.slicing import SliceScorer
 
 scorer = Scorer(metrics=["accuracy", "f1"])
-scorer = SliceScorer(scorer, slice_names)
-scorer.score(
-    S=S_test,
-    golds=Y_test,
-    preds=preds_test,
-    probs=preds_to_probs(preds_test, 2),
-    as_dataframe=True,
+slice_scorer = SliceScorer(scorer, slice_names)
+slice_scorer.score(
+    S=S_test, golds=Y_test, preds=preds_test, probs=probs_test, as_dataframe=True
 )
-
-# %% [markdown]
-# ## 5. Improve slice performance
-#
-# In classification tasks, we might attempt to increase slice performance with techniques like _oversampling_ (i.e. with PyTorch's [`WeightedRandomSampler`](https://pytorch.org/docs/stable/data.html#torch.utils.data.WeightedRandomSampler)).
-# This would shift the training distribution to over-represent certain minority populations.
-# Intuitively, we'd like to show the model more `short_link` examples so that the representation is better suited to handle them.
-#
-# A technique like upsampling might work with a small number of slices, but with hundreds or thousands or production slices, it could quickly become intractable to tune upsampling weights per slice.
-# In the following section, we show a modeling approach that we call _Slice-based Learning,_ which handles numerous slices using with slice-specific representation learning.
 
 # %% [markdown]
 # ### Write additional slicing functions (SFs)
@@ -355,7 +279,7 @@ scorer.score(
 # We'll take inspiration from the labeling tutorial to write additional slicing functions.
 
 # %%
-from snorkel.slicing import SlicingFunction, slicing_function, nlp_slicing_function
+from snorkel.slicing import SlicingFunction, slicing_function
 from snorkel.preprocess import preprocessor
 
 
@@ -374,13 +298,6 @@ def make_keyword_sf(keywords):
 
 keyword_subscribe = make_keyword_sf(keywords=["subscribe"])
 keyword_please = make_keyword_sf(keywords=["please", "plz"])
-
-
-# Leverage @nlp_slicing_function
-@nlp_slicing_function()
-def has_person_nlp(x):
-    """Ham comments mention specific people and are short."""
-    return len(x.doc) < 20 and any([ent.label_ == "PERSON" for ent in x.doc.ents])
 
 
 # Regex-based SF
@@ -412,29 +329,6 @@ def textblob_polarity(x):
     return x.polarity > 0.9
 
 
-extra_sfs = [
-    keyword_subscribe,
-    keyword_please,
-    regex_check_out,
-    short_comment,
-    has_person_nlp,
-    textblob_polarity,
-]
-
-sfs = [short_link] + extra_sfs
-
-# %%
-applier = PandasSFApplier(sfs)
-S_train = applier.apply(df_train)
-S_valid = applier.apply(df_valid)
-S_test = applier.apply(df_test)
-
-# %%
-slice_names = [sf.name for sf in sfs]
-add_slice_labels(dl_train, spam_task, S_train, slice_names)
-add_slice_labels(dl_valid, spam_task, S_valid, slice_names)
-add_slice_labels(dl_test, spam_task, S_test, slice_names)
-
 # %% [markdown]
 # Like we saw above, we'd like to visualize examples in the slice.
 # In this case, most examples with high-polarity sentiments are strong opinions about the video — hence, they are usually relevant to the video, and the corresponding labels are $0$.
@@ -444,12 +338,63 @@ polarity_df = slice_dataframe(df_valid, textblob_polarity)
 polarity_df[["text", "label"]].head()
 
 # %% [markdown]
+# Like we did above, we can evaluate model performance on _all SFs_ using the `SliceScorer`.
+
+# %%
+extra_sfs = [
+    keyword_subscribe,
+    keyword_please,
+    regex_check_out,
+    short_comment,
+    textblob_polarity,
+]
+
+sfs = [short_link] + extra_sfs
+slice_names = [sf.name for sf in sfs]
+
+# %%
+applier = PandasSFApplier(sfs)
+S_test = applier.apply(df_test)
+
+slice_scorer = SliceScorer(scorer, slice_names)
+slice_scorer.score(
+    S=S_test, golds=Y_test, preds=preds_test, probs=probs_test, as_dataframe=True
+)
+
+# %% [markdown]
+# ## 5. Improve slice performance
+#
+# In classification tasks, we might attempt to increase slice performance with techniques like _oversampling_ (i.e. with PyTorch's [`WeightedRandomSampler`](https://pytorch.org/docs/stable/data.html#torch.utils.data.WeightedRandomSampler)).
+# This would shift the training distribution to over-represent certain minority populations.
+# Intuitively, we'd like to show the model more `short_link` examples so that the representation is better suited to handle them.
+#
+# A technique like upsampling might work with a small number of slices, but with hundreds or thousands or production slices, it could quickly become intractable to tune upsampling weights per slice.
+# In the following section, we show a modeling approach that we call _Slice-based Learning,_ which handles numerous slices using with slice-specific representation learning.
+
+# %% [markdown]
 # ### Representation learning with slices
 
 # %% [markdown]
 # To cope with scale, we will attempt to learn and combine many slice-specific representations with an attention mechanism.
 # (For details, please see our technical report — coming soon!)
-# Using the helper, [`convert_to_slice_tasks`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/slicing/snorkel.slicing.convert_to_slice_tasks.html), we initialize slice-specific tasks to learn "expert task heads" for each slice, in the style of multi-task learning.
+
+# %% [markdown]
+# First, we update our existing dataloader with our new slices.
+
+# %%
+applier = PandasSFApplier(sfs)
+S_train = applier.apply(df_train)
+S_valid = applier.apply(df_valid)
+
+# %%
+slice_names = [sf.name for sf in sfs]
+add_slice_labels(dl_train, spam_task, S_train, slice_names)
+add_slice_labels(dl_valid, spam_task, S_valid, slice_names)
+add_slice_labels(dl_test, spam_task, S_test, slice_names)
+
+# %% [markdown]
+# Using the helper, [`convert_to_slice_tasks`](https://snorkel.readthedocs.io/en/redux/packages/_autosummary/slicing/snorkel.slicing.convert_to_slice_tasks.html), we can convert our original `spam_task` into slice-specific tasks.
+# These will be used to learn "expert task heads" for each slice, in the style of multi-task learning.
 # The original `spam_task` now contains the attention mechanism to then combine these slice experts.
 
 # %%
